# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Jobseek dashboard API
  version: 0.1.0
  description: Versioned private dashboard API. All data routes require authentication.
servers:
  - url: /api/v1
paths:
...
  /workflow/roles:
    get:
      operationId: listRoleWorkflows
      summary: List each selected role with its authoritative journey stage
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Selected roles and stages
          content:
            { application/json: { schema: { $ref: '#/components/schemas/RoleWorkflowList' } } }
  /opportunities/{id}/workflow:
    parameters: [{ $ref: '#/components/parameters/RecordId' }]
    get:
      operationId: getRoleWorkflow
      summary: Read one selected role's authoritative journey stage
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Role stage
          content:
            { application/json: { schema: { $ref: '#/components/schemas/RoleWorkflowState' } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    RecordId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jobseek_session
      description: Owner session. Cookie mutations also require X-CSRF-Token and same-origin Origin.
    bearerAuth:
      type: http
      scheme: bearer
      description: Named agent token with operation-specific scopes.
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    RoleWorkflowState:
      type: object
      required: [opportunityId, stage, revision, updatedAt, decisionAt, opportunityRevision]
      properties:
        opportunityId: { type: string }
        stage:
          {
            type: string,
            enum:
              [
                selected,
                checking,
                checked,
                answering,
                answered,
                preparing,
                prepared,
                reviewing,
                sent,
                blocked,
              ],
          }
        revision: { type: integer, format: int64 }
        blockedReason: { type: string }
        updatedAt: { type: string, format: date-time }
        decisionAt: { type: string, format: date-time }
        opportunityRevision: { type: integer, format: int64 }
    RoleWorkflowList:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/RoleWorkflowState' } }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          enum:
            [
              validation_error,
              not_found,
              unauthenticated,
              forbidden,
              csrf_failed,
              rate_limited,
              conflict,
              unavailable,
              internal_error,
            ]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...